src/app/extensions/cam-ahu/pages/camfil-ahu-detail/camfil-ahu-slot-type/camfil-ahu-slot-type.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { provideMockStore } from '@ngrx/store/testing';
import { TranslateModule } from '@ngx-translate/core';

import { CamfilAhuSlotTypeComponent } from './camfil-ahu-slot-type.component';

describe('Camfil Ahu Slot Type Component', () => {
  let component: CamfilAhuSlotTypeComponent;
  let fixture: ComponentFixture<CamfilAhuSlotTypeComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [CamfilAhuSlotTypeComponent],
      imports: [TranslateModule.forRoot()],
      providers: [provideMockStore()],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(CamfilAhuSlotTypeComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
